user controller

+++ crud-app/app/Http/Controllers/ProductController.php
@@ -26,6 +26,11 @@ class ProductController extends Controller
 
     public function create()
     {
+
+        if ($request->user()->cannot('create', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         $product = new Product;
         return view('products.create', ['product' => $product]);
     }
@@ -35,6 +40,10 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        if ($request->user()->cannot('create', Product::class)) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         Product::create($this->validatedData($request));
         return redirect()->route('products.index')->with('success', 'Product was added');
     }
@@ -54,6 +63,10 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        if ($request->user()->cannot('update', [Product::class, Product::findOrFail($id)])) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        }
+
         $product =  Product::findOrFail($id);
         return view('products.edit', ['product' => $product]);
     }
@@ -63,6 +76,10 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if ($request->user()->cannot('update', [Product::class, Product::findOrFail($id)])) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         Product::findOrFail($id)->update($this->validatedData($request));
         return redirect()->route('products.index')->with('success', 'Product was updated');
                              }
@@ -72,6 +89,10 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
+        if ($request->user()->cannot('delete', [Product::class, Product::findOrFail($id)])) {
+            return redirect()->route('products.index')->with('error', 'You do not have permission');
+        };
+
         $product = Product::findOrFail($id);
         $product->delete();
         return redirect()->route('products.index')->with('success', 'Product was deleted');
